# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/questions": {
      "get": {
        "operationId": "AssessmentController_questions",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "question.manage：分页查询授权题库",
        "tags": ["题库、作业、练习、考试与成绩"]
      },
      "post": {
        "operationId": "AssessmentController_createQuestion",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["题库、作业、练习、考试与成绩"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": ["single", "multiple", "boolean", "blank", "short", "composite"]
                  },
                  "stem": {
                    "type": "string",
                    "maxLength": 20000
                  },
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 100
                        },
                        "text": {
                          "type": "string",
                          "maxLength": 4000
                        }
                      },
                      "required": ["id", "text"],
                      "additionalProperties": false
                    },
                    "maxItems": 30,
                    "default": []
                  },
                  "answer": {},
                  "explanation": {
                    "type": "string",
                    "maxLength": 20000,
                    "default": ""
                  },
                  "rules": {
                    "type": "object",
                    "properties": {
                      "partialCredit": {
                        "type": "boolean",
                        "default": false
                      },
                      "caseSensitive": {
                        "type": "boolean",
                        "default": false
                      },
                      "trim": {
                        "type": "boolean",
                        "default": true
                      },
                      "collapseWhitespace": {
                        "type": "boolean",
                        "default": true
                      }
                    },
                    "required": [],
                    "additionalProperties": false,
                    "default": {}
                  },
                  "scoreCents": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100000000
                  },
                  "difficulty": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5,
                    "default": 2
                  },
                  "knowledgePoints": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "maxLength": 100
                    },
                    "maxItems": 30,
                    "default": []
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "maxLength": 100
                    },
                    "maxItems": 30,
                    "default": []
                  },
                  "courseId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "chapterId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "nullable": true
                  },
                  "scope": {
                    "type": "string",
                    "enum": ["private", "shared"],
                    "default": "private"
                  },
                  "practiceEnabled": {
                    "type": "boolean",
                    "default": false
                  },
                  "children": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": ["single", "multiple", "boolean", "blank", "short", "composite"]
                        },
                        "stem": {
                          "type": "string",
                          "maxLength": 20000
                        },
                        "options": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "minLength": 1,
                                "maxLength": 100
                              },
                              "text": {
                                "type": "string",
                                "maxLength": 4000
                              }
                            },
                            "required": ["id", "text"],
                            "additionalProperties": false
                          },
                          "maxItems": 30,
                          "default": []
                        },
                        "answer": {},
                        "explanation": {
                          "type": "string",
                          "maxLength": 20000,
                          "default": ""
                        },
                        "rules": {
                          "type": "object",
                          "properties": {
                            "partialCredit": {
                              "type": "boolean",
                              "default": false
                            },
                            "caseSensitive": {
                              "type": "boolean",
                              "default": false
                            },
                            "trim": {
                              "type": "boolean",
                              "default": true
                            },
                            "collapseWhitespace": {
                              "type": "boolean",
                              "default": true
                            }
                          },
                          "required": [],
                          "additionalProperties": false,
                          "default": {}
                        },
                        "scoreCents": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 100000000
                        },
                        "difficulty": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 5,
                          "default": 2
                        },
                        "knowledgePoints": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "maxLength": 100
                          },
                          "maxItems": 30,
                          "default": []
                        },
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "maxLength": 100
                          },
                          "maxItems": 30,
                          "default": []
                        },
                        "id": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 100
                        }
                      },
                      "required": ["type", "stem", "answer", "scoreCents", "id"],
                      "additionalProperties": false
                    },
                    "maxItems": 30,
                    "default": []
                  }
                },
                "required": ["type", "stem", "answer", "scoreCents", "courseId"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/api/questions/export": {
      "get": {
        "operationId": "AssessmentController_exportQuestions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["题库、作业、练习、考试与成绩"]
      }
    },
    "/api/questions/template": {
      "get": {
        "operationId": "AssessmentController_template",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["题库、作业、练习、考试与成绩"]
      }
    },
...
    "/api/questions/{id}/versions": {
      "get": {
        "operationId": "AssessmentController_versions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["题库、作业、练习、考试与成绩"]
      }
    },
    "/api/questions/{id}/copy": {
      "post": {
        "operationId": "AssessmentController_copy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["题库、作业、练习、考试与成绩"]
      }
    },
    "/api/questions/{id}/favorite": {
      "put": {
        "operationId": "AssessmentController_favorite",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["题库、作业、练习、考试与成绩"]
      }
    },
    "/api/papers": {
      "get": {
        "operationId": "AssessmentController_papers",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["题库、作业、练习、考试与成绩"]
      },
      "post": {
        "operationId": "AssessmentController_createPaper",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["题库、作业、练习、考试与成绩"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "courseId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "title": {},
                  "questionVersionIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    },
                    "maxItems": 200,
                    "default": []
                  },
                  "rule": {
                    "type": "object",
                    "properties": {
                      "count": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 100
                      },
                      "type": {
                        "type": "string"
                      },
                      "difficulty": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 5
                      },
                      "knowledgePoint": {
                        "type": "string",
                        "maxLength": 100
                      },
                      "practiceEnabled": {
                        "type": "boolean"
                      }
                    },
                    "required": ["count"],
                    "additionalProperties": false
                  }
                },
                "required": ["courseId", "title"],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/api/assignments/{id}/publish": {
      "post": {
        "operationId": "AssessmentController_publishAssignment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["题库、作业、练习、考试与成绩"]
      }
    },
...
    "/api/assignments/{id}/submissions": {
      "get": {
        "operationId": "AssessmentController_submissions",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["题库、作业、练习、考试与成绩"]
      }
    },
    "/api/assignments/{id}/roster": {
      "get": {
        "operationId": "AssessmentController_assignmentRoster",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["题库、作业、练习、考试与成绩"]
      }
    },
    "/api/assignments/{id}/release": {
      "post": {
        "operationId": "AssessmentController_releaseAssignment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "字段验证失败"
          },
          "401": {
            "description": "未登录或会话已失效"
          },
          "403": {
            "description": "功能、机构、角色上下文或资源范围无权限"
          },
          "404": {
            "description": "资源不存在"
          },
          "409": {
            "description": "版本冲突或重复约束"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": ["题库、作业、练习、考试与成绩"]
      }
    },
...
  },
  "info": {
    "title": "知学 · 学习管理系统 API",
    "description": "同源 Cookie 会话。写请求必须携带登录响应中的 x-csrf-token。所有教学资源逐请求执行机构、当前角色、课程成员资格和对象归属校验。分页 page/pageSize，错误 {error:{code,message,fields,requestId}}。详细参数及示例见 docs/*-api.md。",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "lms_session"
      }
    },
...
  }
}